+++ HISTORY.md
@@ -1,2 +1,10 @@
 ##v0.1.0
 Initial version. Support send IPMI command to BMC via KCS directly under Linux or DOS.
+
+###v0.1.1
+Based on v0.1.0. Add some customized commands for testbed using.
+
+Commands: 
+* `devid`:  check the existence of BMC.
+* `fwinfo`: get build time and revision of the current firmware. 
+